+++ src/main/java/com/neoris/turnosrotativos/repositories/EmpleadoRepository.java
@@ -4,6 +4,6 @@ import com.neoris.turnosrotativos.entities.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
-    public Empleado findEmpleadoByNroDocumento(int nroDocumento);
+    public Empleado findEmpleadoByNroDocumento(Long nroDocumento);
     public Empleado findEmpleadoByEmail(String email);
 }
